+++ Kros.MassTransit.AzureServiceBus/Endpoints/Endpoint.cs
@@ -36,7 +36,7 @@ namespace Kros.MassTransit.AzureServiceBus.Endpoints
         /// <typeparam name="TConsumer">Type of message consumer.</typeparam>
         /// <param name="provider">Service provider (DI container).</param>
         /// <param name="configure">Delegate to configure consumer.</param>
-        public abstract void AddConsumerWithDependencies<TConsumer>(
+        public abstract void AddConsumer<TConsumer>(
             IServiceProvider provider,
             Action<IConsumerConfigurator<TConsumer>> configure = null) where TConsumer : class, IConsumer;
